[fix #186] Document 구조 변경
