+++ src/main/resources/elastic/member-mappings.json
@@ -10,9 +10,6 @@
     "email": {
       "type": "text",
       "analyzer": "member_search_analyzer"
-    },
-    "image": {
-      "type": "text"
     }
   }
 }
